requirements.txt
streamlit
pandas
python-dotenv
supabase
xhtml2pdf